Build 20150516-001
+ Changed all absolute path to relative path
+ Added the possibility to show success and warning message when
  submit data on administration page
+ Fixed error on graph page
+ Removed fixed page title and get it from database

+++ application/controllers/operador.php
@@ -62,7 +62,8 @@ class Operador extends CI_Controller {
 
   public function agregar($sector_url) {
     // Buffer para almacenar mensaje
-    $msg = '';
+    $successMsg = '';
+    $warningMsg = '';
 
     $sector_data = $this->planillas->get_sector_data_by_url($sector_url);
 
@@ -70,7 +71,6 @@ class Operador extends CI_Controller {
     if ($inputs) {
       // Buffer de los datos para insertar
       $datos = array();
-
 
       // Loop de los inputs
       foreach ($inputs as $input_maquina => $input_valor) {
@@ -86,27 +86,12 @@ class Operador extends CI_Controller {
       }
 
       // Insertar datos en conjunto
-      if ('access_error' != $this->planilla_datos->insert($this->session->userdata('id'), $datos)) {
+      if ($this->planilla_datos->insert($this->session->userdata('id'), $datos)) {
         // Configurar mensaje y redireccionar
-        $msg = 'La información ha sido ingresada correctamente';
+        $successMsg = 'La información ha sido ingresada correctamente';
       } else {
-        $msg = '¡Detectamos intento de acceso de datos no correspondiente!';
+        $warningMsg = '¡Detectamos intento de acceso de datos no correspondiente!';
       }
-    }
-
-    switch ($sector_url) {
-      case 'cocedor':
-        $titulo = 'Agregar Temperatura';
-        break;
-      case 'prensa':
-        $titulo = 'Agregar Presión hidráulica (BAR)';
-        break;
-      case 'evaporador':
-        $titulo = 'Agregar % Solido Concentrado';
-        break;
-      case 'secador':
-        $titulo = 'Agregar % Fluido Termico';
-        break;
     }
 
     // Buffer de inputs
@@ -116,7 +101,8 @@ class Operador extends CI_Controller {
          $maquinasLength = sizeof($maquinas); $i < $maquinasLength; ++$i) {
       array_push($inputs, array(
         'nombre' => $maquinas[$i]->nombre, 
-        'maquina' => $maquinas[$i]->id
+        'maquina' => $maquinas[$i]->id, 
+        'valor' => 0
       ));
     }
 
@@ -127,24 +113,33 @@ class Operador extends CI_Controller {
         'enlace_base_planilla' => base_url('planilla/'), 
         'sectores' => $this->planillas->get_all_sector(), 
         'nombre' => $this->session->userdata('nombre'),
-        'msg' => $msg
+        'successMsg' => $successMsg, 
+        'warningMsg' => $warningMsg
       ));
     }
     $this->load->view('operador/modificar', array(
-      'titulo' => $titulo, 
+      'titulo' => 'Agregar ' . $sector_data->medida, 
       'actionURL' => base_url('/operador/'. $sector_url .'/agregar'),
       'inputs' => $inputs
     ));
   }
 
-  public function editar($sector_url, $dato_id) {
+  public function editar($sector_url, $dato_id = 0) {
     // Prevenir ser accedido por no administradores
     if (!$this->session->userdata('isAdmin')) {
       redirect('/');
     }
 
+    // Evitar de haber ingresado sin el id del dato
+    if (!$dato_id) {
+      redirect('/planilla/' . $sector_url);
+    }
+
+    $sector_data = $this->planillas->get_sector_data_by_url($sector_url);
+
     // Buffer de mensaje
-    $msg = '';
+    $successMsg = '';
+    $warningMsg = '';
 
     // Ver si hay envío de datos
     $inputs = $this->input->post('value');
@@ -154,36 +149,24 @@ class Operador extends CI_Controller {
       // Si el valor de input no es vacio y es numerico
       if (!empty($inputs[$maquina_id]) && is_numeric($inputs[$maquina_id])) {
         // Actualizar
-        $this->planilla_datos->update($id, $maquina_id, $inputs[$maquina_id]);
+        if ($this->planilla_datos->update($dato_id, $maquina_id, $inputs[$maquina_id])) {
+          $this->session->set_flashdata('successMsg','El registro ha sido modificado correctamente');
+        } else {
+          $this->session->set_flashdata('warningMsg','El registro no ha podido ser modificado');
+        }
 
         // Redireccionar
-        $this->session->set_flashdata('msg','El registro ha sido modificado correctamente');
-        redirect('planilla/' . $sector_url . '/view/' . $planilla_id);
+        redirect('/planilla/' . $sector_url . '/ver/' . $planilla_id);
       } else {
         // Mostrar error
-        $msg = 'El dato ingresado es incorrecto';
+        $warningMsg = 'El dato ingresado es incorrecto';
       }
     }
 
-    switch ($sector_url) {
-      case 'cocedor':
-        $titulo = 'Editar Temperatura';
-        break;
-      case 'prensa':
-        $titulo = 'Editar Presión hidráulica (BAR)';
-        break;
-      case 'evaporador':
-        $titulo = 'Editar % Solido Concentrado';
-        break;
-      case 'secador':
-        $titulo = 'Editar % Fluido Termico';
-        break;
-    }
-
     // Obtener datos de la maquina
-    $maquina_data = $this->planillas->get_maquina_by_id(
-      $this->planilla_datos->get_by_id($dato_id)->maquina_id
-    );
+    $maquina_dato = $this->planilla_datos->get_by_id($dato_id);
+    $maquina_id = $maquina_dato->maquina_id;
+    $maquina_data = $this->planillas->get_maquina_by_id($maquina_id);
 
     // Mostrar interfaz
     $this->load->view('header');
@@ -191,52 +174,49 @@ class Operador extends CI_Controller {
       'enlace_base_planilla' => base_url('planilla/'), 
       'sectores' => $this->planillas->get_all_sector(), 
       'nombre' => $this->session->userdata('nombre'), 
-      'msg' => $msg
+      'successMsg' => $successMsg,
+      'warningMsg' => $warningMsg
     ));
     $this->load->view('operador/modificar', array(
-      'titulo' => $titulo, 
+      'titulo' => 'Editar ' . $sector_data->medida, 
       'actionURL' => base_url('/operador/'. $sector_url .'/editar/' . $dato_id),
       'inputs' => array(
         array(
           'nombre' => $maquina_data->nombre, 
-          'maquina' => $maquina_id
+          'maquina' => $maquina_id, 
+          'valor' => $maquina_dato->valor, 
         )
       )
     ));
   }
 
-  public function eliminar($sector_url, $dato_id) {
+  public function eliminar($sector_url, $dato_id = 0) {
     // Prevenir ser accedido por no administradores
     if (!$this->session->userdata('isAdmin')) {
       redirect('/');
     }
 
-    $this->planilla_datos->delete($dato_id);
-    redirect('planilla/' . $sector_url);
+    // Evitar de haber ingresado sin el id del dato
+    if ($dato_id) {
+      $this->planilla_datos->delete($dato_id);
+    }
+
+    redirect('/planilla/' . $sector_url);
   }
 
-  public function grafico($sector_url, $planilla_id) {
+  public function grafico($sector_url, $planilla_id = 0) {
+    // Evitar de haber ingresado sin el id de la planilla
+    if (!$planilla_id) {
+      redirect('/');
+    }
+
     // Buffer de mensaje
     $msg = '';
 
-    switch ($sector_url) {
-      case 'cocedor':
-        $titulo = 'Temperatura de Licor';
-        break;
-      case 'prensa':
-        $titulo = 'Presión hidráulica (BAR)';
-        break;
-      case 'evaporador':
-        $titulo = '% Solido Concentrado';
-        break;
-      case 'secador':
-        $titulo = '% Fluido Termico';
-        break;
-    }
+    $sector_data = $this->planillas->get_sector_data_by_url($sector_url);
 
     $planilla_data = $this->planillas->get_by_id($planilla_id);
     $planilla_datos = $this->planilla_datos->get_by_planilla($planilla_id);
-    $titulo .= ' ' . $planilla_data->fecha;
 
     $datos = array();
     for ($i = 0, $planillaDatosLength = sizeof($planilla_datos), $maquinaId; $i < $planillaDatosLength; ++$i) {
@@ -244,13 +224,16 @@ class Operador extends CI_Controller {
 
       if (!isset($datos[$maquinaId]['min'])) {
         $maquina_dato = $this->planillas->get_maquina_by_id($maquinaId);
+        $datos[$maquinaId]['nombre'] = $maquina_dato->nombre;
         $datos[$maquinaId]['min'] = $maquina_dato->min;
         $datos[$maquinaId]['max'] = $maquina_dato->max;
         $datos[$maquinaId]['unidad'] = $maquina_dato->unidad;
+        $datos[$maquinaId]['tiempo'] = array();
+        $datos[$maquinaId]['valor'] = array();
       }
 
-      array_push($datos[$maquinaId]['tiempo'], $datos[$planilla_datos[$i]]->tiempo);
-      array_push($datos[$maquinaId]['valor'], $datos[$planilla_datos[$i]]->valor);
+      array_push($datos[$maquinaId]['tiempo'], $planilla_datos[$i]->tiempo);
+      array_push($datos[$maquinaId]['valor'], $planilla_datos[$i]->valor);
     }
 
     $this->load->view('header');
@@ -263,15 +246,21 @@ class Operador extends CI_Controller {
       ));
     }
     $this->load->view('/operador/grafico', array(
-      'titulo' => $titulo,
+      'titulo' => $sector_data->medida, 
+      'fecha' => $planilla_data->fecha,
       'datos' => $datos,
     ));
   }
     
-  public function pdf($sector_url, $planilla_id) {
+  public function pdf($sector_url, $planilla_id = 0) {
     // Prevenir ser accedido por no administradores
     if (!$this->session->userdata('isAdmin')) {
       redirect('/');
+    }
+
+    // Evitar de haber ingresado sin el id de la planilla
+    if (!$planilla_id) {
+      redirect('/planilla/' . $sector_url);
     }
 
 /*
